migrate non pages components: /home
