+++ src/components/Home/BuiltForSection/BuiltForBlobAndImage.tsx
