+++ src/components/Home/ConclusionSection/ConclusionSection.tsx
@@ -1,9 +1,9 @@
 import { Trans } from '@lingui/macro'
 import { Button } from 'antd'
+import { TickIconList } from 'components/Home/ConclusionSection/TickIconList'
+import { SectionContainer } from 'components/Home/SectionContainer'
 import Image from 'next/image'
 import Link from 'next/link'
-import { SectionContainer } from '../SectionContainer'
-import { TickIconList } from './TickIconList'
 
 export function ConclusionSection() {
   return (
