+++ src/components/Home/ConclusionSection/TickIconList.tsx
@@ -1,5 +1,5 @@
 import { t } from '@lingui/macro'
-import { TickIconListItem } from '../../../components/TickIconListItem'
+import { TickIconListItem } from '../../TickIconListItem'
 
 export function TickIconList() {
   return (
