+++ src/components/Home/FaqList/index.ts
