+++ src/components/Home/HomepageProjectCard.tsx
@@ -1,11 +1,11 @@
 import { Trans } from '@lingui/macro'
 import { Skeleton } from 'antd'
+import { HomepageCard } from 'components/Home/HomepageCard'
 import ProjectLogo from 'components/ProjectLogo'
 import ETHAmount from 'components/currency/ETHAmount'
 import { useProjectMetadata } from 'hooks/useProjectMetadata'
 import { DBProject } from 'models/dbProject'
 import { v2v3ProjectRoute } from 'utils/routes'
-import { HomepageCard } from './HomepageCard'
 
 function Statistic({
   name,
