+++ src/components/Home/JuicyPicksSection/constants.ts
