+++ src/components/Home/JuicyPicksSection/hooks/useExploreCategories.ts
