+++ src/components/Home/JuicyPicksSection/index.ts
