+++ src/components/Home/StatsSection.tsx
