+++ src/components/Home/WhyJuiceboxSection/index.ts
@@ -1,0 +1,1 @@
+export { WhyJuiceboxSection } from 'components/Home/WhyJuiceboxSection/WhyJuiceboxSection'
