--- src/pages/home/WhyJuiceboxSection/index.ts
@@ -1,1 +1,0 @@
-export { WhyJuiceboxSection } from './WhyJuiceboxSection'
